refactor: mv attention to core/layers
